# static/openapi_skg-if_cessda.yaml
openapi: 3.1.0
info:
  version: 1.0.0
  title: SKG-IF OpenAPI - compatible with SKG-IF Data Model 1.1.0
  termsOfService: 'https://skg-if.github.io/interoperability-framework/'
  contact:
    url: 'https://github.com/skg-if'
  license:
    name: CC-BY
    url: 'http://creativecommons.org/licenses/by/4.0/'
  x-logo:
    url: 'https://www.rd-alliance.org/wp-content/uploads/2024/04/RDA_Logotype_CMYK.png'
  description: |
    Below the line is the original SKG-IF OpenAPI documentation that has only been modified to reflect the implementation status of CESSDA's endpoints. Endpoints that are not available have a text saying **"Not yet implemented"** next to the endpoint name. Filters that are not applicable / not implemented are marked with a strikethrough.

    --------------------------------------

    OpenAPI end points specifications for [SKG-IF](https://skg-if.github.io/interoperability-framework/)

    Last version of the SKG-IF OpenAPI is available at: https://skg-if.github.io/api/

    # Note for implementers :
      - To validate a live implementation server see: [.github/WORFLOW.md](https://github.com/skg-if/api/blob/main/.github/WORFLOW.md)
      - Please refer to the [SKG-IF OpenAPI Implementer documentation](https://docs.google.com/document/d/1t7b7h28UTtM56Sda4NGJIp0hnQfGbcVVGn12fny9wfI/edit?usp=sharing)
      - [RDA SKG-IF WG](https://skg-if.github.io/) does not provide any official implementation of the current specification.
      - Recommended OpenAPI viewers:
...
paths:
...
  '/organisations/{local_identifier}':
    get:
      tags:
          - Organisation
      summary: Get organisation by id
      description: |
          Get `organisation` by id.  See definition in SKG-IF [Agent](https://skg-if.github.io/interoperability-framework/docs/agent.html) ( entity_type:organisation) .\
      operationId: getOrganisationById
      parameters:
        - $ref : '#/components/parameters/localIdPathParam'
      responses:
        '200':
          description: Success
          content:
            # not 'application/json-ld' to be compatible with StopLight PRISM tool
            application/json:
              schema:
                required: ['@context','@graph']
                properties:
                  "@context":
                    type: array
                    description: JSON-LD context # not displayed in Stoplight elements
                    minItems: 3 # http://../skg-if.json, http://../skg-if-api.json and @base
                    uniqueItems: true
                    items:
                      anyOf:
                        - $ref: '#/components/schemas/JsonLdCtxDataModel'
                        - $ref: '#/components/schemas/JsonLdCtxSearch'
                        - $ref: '#/components/schemas/JsonLdCtxBaseOrMore'
                    examples:
                      - ["https://w3id.org/skg-if/context/1.1.0/skg-if.json", "https://w3id.org/skg-if/context/1.0.0/skg-if-api.json", {"@base" : "https://w3id.org/skg-if/sandbox/acme/"}]
                  meta:
                    type: object
                    $ref: '#/components/schemas/MetaSingleEntity'
                  "@graph":
                    type: array
                    minItems: 1
                    maxItems: 1
                    items:
                      $ref: "#/components/schemas/Organisation"
              examples:
                OrgBrownUniversity:
                  summary: "Brown University"
                  value :
                    "@context":
                      - "https://w3id.org/skg-if/context/1.1.0/skg-if.json"
                      - "https://w3id.org/skg-if/context/1.0.0/skg-if-api.json"
                      - "@base" : "https://w3id.org/skg-if/sandbox/acme/"
                    "@graph":
                      - local_identifier: organisation-2-bu
                        entity_type: 'organisation'
                        identifiers:
                          - scheme: ror
                            value: https://ror.org/05gq02987
                        name: Brown University.
                        short_name: BU
                        country: US
                        types:
                          - education
                OrgJohnCarterBrownLibrary:
                  summary: "John Carter Brown Library - Brown University"
                  value :
                    "@context":
                      - "https://w3id.org/skg-if/context/1.1.0/skg-if.json"
                      - "https://w3id.org/skg-if/context/1.0.0/skg-if-api.json"
                      - "@base" : "https://w3id.org/skg-if/sandbox/acme/"
                    "@graph":
                      - local_identifier: organisation-3-jcbl
                        entity_type: 'organisation'
                        identifiers:
                          - scheme: ror
                            value: https://ror.org/05gq02987
                          - scheme: ror
                            value: https://ror.org/0274ane14
                        name: John Carter Brown Library, Brown University, Providence RI, US
                        short_name: BU
                        country: US
                        types:
                          - archive
                OrgLabScienceNoRor:
                  summary: "Lab of science"
                  value :
                    "@context":
                      - "https://w3id.org/skg-if/context/1.1.0/skg-if.json"
                      - "https://w3id.org/skg-if/context/1.0.0/skg-if-api.json"
                      - "@base" : "https://w3id.org/skg-if/sandbox/acme/"
                    "@graph":
                      - local_identifier: organisation-5-lsa
                        entity_type: 'organisation'
                        name: Lab of science, unknown University, Antarctica
                OrgEuropeanCommission:
                  summary: "European Commission Funder"
                  value :
                    "@context":
                      - "https://w3id.org/skg-if/context/1.1.0/skg-if.json"
                      - "https://w3id.org/skg-if/context/1.0.0/skg-if-api.json"
                      - "@base" : "https://w3id.org/skg-if/sandbox/acme/"
                    "@graph":
                      - local_identifier: organisation-6-eu
                        entity_type: 'organisation'
                        identifiers:
                          - scheme: ror
                            value: https://ror.org/00k4n6c32
                        name: "European Commission"
                        short_name: EU
                        types:
                          - funder
        '404':
          $ref: '#/components/responses/NotFoundEntity'
  '/organisations':
    get:
      tags:
          - Organisation
      summary: Get list of organisation.
      operationId: getOrganisations
      description: |
          Get a list of `organisation`. See definition in SKG-IF [Agent](https://skg-if.github.io/interoperability-framework/docs/agent.html) ( entity_type:organisation) .

          __Attribute filters__ \
          You can filter using these attributes of the `organisation` object
          | Filter key| Attribute filter example |
          | -------- | -------- |
          | identifiers.id | `identifiers.id:05gq02987` |
          | identifiers.scheme | `identifiers.scheme:ror` |
          | name | `name:Brown University` |
          | short_name |  `short_name:BU` |
          | website | `website:https://www.brown.edu` |
          | <s>country</s> | `country:US` |

          __Convenience filters__ \
          These filters (`cf.*`) aren't attributes of the `organization` object, but they're handy for solving some common use cases.

          | Filter key | Filter value | List returned | Attribute filter example |
          | -------- | -------- | ------- | ------- |
          | cf.search.name | a string | `organisation` with name containing the filter value | `cf.search.name:Brown` |

          __Query parameter syntax examples__ \

           * `/organisations?filter=identifiers.scheme:ror,identifiers.id:05gq02987`
           * `/organisations?filter=cf.search.name:Brown`
           * `/organisations?filter=cf.search.name:Brown&page=1&page_size=5`

      parameters:
        - name: filter
          in: query
          description: filter. Coma separated filter_name:filter_value elements ( filter_name_1:filter_value_1,filter_name_2:filter_value_2,filter_name_3:filter_value_3... ). Server side operator used is _AND_.
          examples :
            rorid_ex:
              value: identifiers.scheme:ror,identifiers.id:05gq02987
              summary : Get all organisation with identifier scheme 'ror' and id '05gq02987'
            name_ex:
              value: cf.search.name:Brown
              summary : Get all organisation with name containing 'Brown'
          schema:
            type: string
            pattern: '^(,?.+:.+)*$'
        - $ref : '#/components/parameters/pageQueryParam'
        - $ref : '#/components/parameters/pageSizeQueryParam'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                properties:
                  "@context":
                    type: array
                    description: JSON-LD context # not displayed in Stoplight elements
                    minItems: 3 # http://../skg-if.json, http://../skg-if-api.json and @base
                    uniqueItems: true
                    items:
                      anyOf:
                        - $ref: '#/components/schemas/JsonLdCtxDataModel'
                        - $ref: '#/components/schemas/JsonLdCtxSearch'
                        - $ref: '#/components/schemas/JsonLdCtxBaseOrMore'
                    examples:
                      - ["https://w3id.org/skg-if/context/1.1.0/skg-if.json", "https://w3id.org/skg-if/context/1.0.0/skg-if-api.json", {"@base" : "https://w3id.org/skg-if/sandbox/acme/"}]
                  meta:
                    type: object
                    $ref: '#/components/schemas/MetaSearch'
                  "@graph":
                    type: array
                    items:
                      oneOf:
                        - $ref: '#/components/schemas/Organisation'
                        - $ref: '#/components/schemas/LocalIdentifierRef'
                required: [ "@context", "meta", "@graph" ]
        '404':
          $ref: '#/components/responses/NotFoundSearch'
...
components:
  securitySchemes:
    main_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: 'http://example.com/api/oauth/dialog'
          scopes:
            'read:users': read users info
            'write:users': modify or remove users
    api_key:
      type: apiKey
      in: header
      name: api_key
    basic_auth:
      type: http
      scheme: basic
  parameters:
    localIdPathParam:
      name: local_identifier
      in: path
      description: |
        Entity id.
        - Example entity id with as URL : `http://skg-base-domain.com/prod-c66c6-38be-4d5f-85db-d44c9`. To define your SKG base domain ( `@base` ), please refer to the [SKG-IF API doc](https://skg-if.github.io/api/)
        - Example entity id without its SKG base domain : `prod-c66c6-38be-4d5f-85db-d44c9`

        The entity endpoint MUST resolve ids as full URL and SHOULD resolve ids without SKG base domain.
      required: true
      schema:
        type: string
#    filterQueryParam: # filterQueryParam duplicated in each search endpoint ( only way to add distinct examples for each endpoint )
#      name: filter
#      in: query
#      description: |
#            Search filter. Format : Coma separated filter_name:filter_value elements ( filter_name_1:filter_value_1,filter_name_2:filter_value_2,filter_name_3:filter_value_3... ). Server side operator used is _AND_.
#      schema:
#        type: string
#        pattern: '^(,?.+:.+)*$'
    pageQueryParam:
      name: page
      in: query
      description: |
            Search page reference (page number or search iteration token).\
            Note: As a client application you should build your iterating logic using the `meta.next_page.local_identifier` and `meta.prev_page.local_identifier` URLs
      schema:
        type: string
    pageSizeQueryParam:
      name: page_size
      in: query
      description: |
            Search page size. limit search result size.
      schema:
        type: integer
  responses:
    NotFoundEntity:
      description: |
        No entity found for the given id. The `@graph` array is empty.
      content:
        application/json:
          schema:
            required: ['@context','@graph']
            properties:
              "@context":
                type: array
                description: JSON-LD context # not displayed in Stoplight elements
                minItems: 3 # http://../skg-if.json, http://../skg-if-api.json and @base
                uniqueItems: true
                items:
                  anyOf:
                    - $ref: '#/components/schemas/JsonLdCtxDataModel'
                    - $ref: '#/components/schemas/JsonLdCtxSearch'
                    - $ref: '#/components/schemas/JsonLdCtxBaseOrMore'
                examples:
                  - ["https://w3id.org/skg-if/context/1.1.0/skg-if.json", "https://w3id.org/skg-if/context/1.0.0/skg-if-api.json", {"@base" : "https://w3id.org/skg-if/sandbox/acme/"}]
              meta:
                type: object
                $ref: '#/components/schemas/MetaSingleEntity'
              "@graph":
                type: array
                maxItems: 0
                items: {}
          examples:
            NotFoundEntityEx01:
              summary: No entity found for the given id
              value:
                "@context": [
                  "https://w3id.org/skg-if/context/1.1.0/skg-if.json",
                  "https://w3id.org/skg-if/context/1.0.0/skg-if-api.json",
                  {"@base" : "https://w3id.org/skg-if/sandbox/acme/"}
                ]
                "meta":
                  local_identifier: https://example.com/skg-if/api/products/unknown-id
                  entity_type: "single_entity"
                "@graph": []
    NotFoundSearch:
      description: |
        No entity found matching the given filter. The `@graph` array is empty.
      content:
        application/json:
          schema:
            properties:
              "@context":
                type: array
                description: JSON-LD context # not displayed in Stoplight elements
                minItems: 3 # http://../skg-if.json, http://../skg-if-api.json and @base
                uniqueItems: true
                items:
                  anyOf:
                    - $ref: '#/components/schemas/JsonLdCtxDataModel'
                    - $ref: '#/components/schemas/JsonLdCtxSearch'
                    - $ref: '#/components/schemas/JsonLdCtxBaseOrMore'
                examples:
                  - ["https://w3id.org/skg-if/context/1.1.0/skg-if.json", "https://w3id.org/skg-if/context/1.0.0/skg-if-api.json", {"@base" : "https://w3id.org/skg-if/sandbox/acme/"}]
              meta:
                type: object
                $ref: '#/components/schemas/MetaSearch'
              "@graph":
                type: array
                maxItems: 0
                items: {}
            required: [ "@context", "meta", "@graph" ]
          examples:
            NotFoundSearchEx01:
              summary: No entity found matching the given filter
              value:
                "@context": [
                  "https://w3id.org/skg-if/context/1.1.0/skg-if.json",
                  "https://w3id.org/skg-if/context/1.0.0/skg-if-api.json",
                  {"@base" : "https://w3id.org/skg-if/sandbox/acme/"}
                ]
                "meta":
                  local_identifier: https://example.com/skg-if/api/products?filter=cf.search.title:unknown
                  entity_type: "search_result_page"
                "@graph": []
  schemas:
    LocalIdentifierRef:
      type: string
      description: |
        A local identifier reference.

        You can pass a local identifier as a string and not a full object entity structure, but it is not recommended.
        ex:
        ```
          "foo":"product-1"
        ```

        This feature is included in the current SKG-IF OpenAPI specifications to remain compatible with SKG-IF model documentation and raw exports.
        This approach creates important limitations for the client applications explained below.

        Instead, we recommend to pass one of the object structure entities compatible with this property, and also fill its properties.
        One additional benefit of object structure is to limit the client API calls to resolve entities and sub-entities.
        ex:
        ```
          "foo": {
            "local_identifier":"product-1",
            "entity_type:product",
            "title":"xxx",
            ...
            "contributions" : [
              {
               "by": {
                 "local_identifier":"person-1",
                 "entity_type:person",
                 "name":"John Doe",
                 ...
               }
              }
            ]
          }
        ```

        This is also very useful for client using your API for autocomplete features.
        A typical autocomplete use case is : search products by term and get a list of products with the contributors names in a single API call.
    ApiItem:
      type: object
      title: 'Link'
      description:  |
         Result element
      required: [
         "local_identifier"
      ]
      properties:
        local_identifier:
          type: string
          description: |
            Unique local identifier as URL
          examples:
            - http://w3id.org/skg-if/sandbox/acme/product-1
            - http://w3id.org/skg-if/sandbox/acme/person-1
        urls:
          type: [array, "null"]
          description: |
            List of API URLs to access to the entity metadata
          items:
            $ref: '#/components/schemas/Link'
...
    Organisation:
      type: object
      title: "Organisation"
      description:  "`organisation` object.  See definition in SKG-IF [Agent](https://skg-if.github.io/interoperability-framework/docs/agent.html) ( entity_type:organisation )."
      allOf:
        - $ref: "#/components/schemas/Entity"
        - type: object
          required: [
            "local_identifier","entity_type"
          ]
          properties :
            entity_type:
              type: string
              default: "organisation"
              pattern: '^organisation$'
              x-faker:
                fake: ["organisation"]
            identifiers:
              description: Objects representing external identifiers for the entity. See [Agent](https://skg-if.github.io/interoperability-framework/docs/agent.html#identifiers) documentation for supported schemes
              type: [array, "null"]
              items:
                required: [ "scheme", "value" ]
                properties:
                  scheme:
                    description: The scheme for the external identifier.
                    type: string
                    x-faker:
                      fake: ["ror"]
                    examples:
                      - 'ror'
                  value:
                    type: string
                    description: "The external identifier."
                    x-faker:
                      fake : ['{{random.alphaNumeric(10)}}']
                    examples:
                      - "https://ror.org/05gq02987"
            name:
              type: [string, "null"]
              x-faker: company.companyName
              examples:
                - "Brown University"
            other_names:
              description:  "Other names, perhaps in different languages, identifying an `organisation`."
              type: [array, "null"]
              items:
                type: string
              examples :
                - ["Brownie"]
            short_name:
              type: [string, "null"]
              #x-faker: company.bsAdjective
              x-faker:
                random.alpha: [{count: 4, upcase: true}]
              examples:
                - "BU"
            website:
              type: [string, "null"]
              description: The website URL for an Organisation
              examples:
                - "https://www.brown.edu/"
            country:
              type: [string, "null"]
              description: The country code expressed as [ISO 3166-1 alpha-2](https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2).
              pattern: '^[A-Z]{2}$'
              examples:
                - "US"
            types:
              description:  "The types of an `organisation`."
              type: [array, "null"]
              items:
                type: string
                enum:
                  - archive
                  - company
                  - education
                  - facility
                  - government
                  - healthcare
                  - nonprofit
                  - funder
                  - research
                  - unspecified
      examples:
        - local_identifier: 'organisation-2-bu'
          entity_type: 'organisation'
          identifiers:
            - "scheme": ror
              "value": https://ror.org/05gq02987
          name: "Brown University"
          short_name: BU
          country: US
          types:
            - education
        - local_identifier: organisation-3-jcbl
          entity_type: 'organisation'
          identifiers:
            - scheme: ror
              value: https://ror.org/05gq02987
            - scheme: ror
              value: https://ror.org/0274ane14
          name: John Carter Brown Library, Brown University, Providence RI, US
          short_name: BU
          country: US
          types:
            - archive
        - local_identifier: organisation-5-lsa
          entity_type: 'organisation'
          name: Lab of science, unknown University, Antarctica
        - local_identifier: organisation-6-eu
          entity_type: 'organisation'
          identifiers:
            - scheme: ror
              value: https://ror.org/00k4n6c32
          name: "European Commission"
          short_name: EU
          types:
            - funder
...
    Entity:
      type: object
      description: "Entity common fields used for entities : `product`,`person`,`organisation`,`grant`,`venue`,`datasource`"
      properties:
        local_identifier:
          type: string
          description: |
            Unique local identifier as URL.
            format : `https://[your-@base-url]/{identifier string}`
            example: `https://w3id.org/skg-if/sandbox/acme/prod-1`

            The local identifier value can contain only the `{identifier string}` part only if the `@base` is present in the `@context` output. \
            The local identifier value can also be an URL reference not included in your `@base` domain.
            This can be useful for entities identified in your system by their external PID only (RoR, DOI...)

            If the entity does not have any stable identifier, an on-the-fly identifier is returned : `otf___<session identifier>___<identifier string>`.
            -  `otf` stands for on-the-fly to explicitly clarify the local identifiers it has been created for the purpose of creating this SKG-IF document;
            -  `<session identifier>` is a string portion that enables the source to uniquely identifier the session in which such SKG-IF document has been created - e.g. it could be the current time of the software run to create the SKG expressed in milliseconds;
            -  `<identifier string>` is a string portion that identifies the particular entity in consideration.
            For instance, a possible example of such a local identifier would be : `otf___1730027051396___person-1`.

            An on-the-fly may not resolve with its `self` API link URL ex: `https://example.com/skg-if/api/persons/otf___1730027051396___person-1`.
            on-the-fly means that this local identifier cannot be resolved outside the session context.
          x-faker:
            fake : ['{{random.uuid}}']
          examples:
            - http://w3id.org/skg-if/sandbox/acme/product-1
            - http://w3id.org/skg-if/sandbox/acme/person-1
            - http://w3id.org/skg-if/sandbox/acme/otf___1730027051396___person-1
            - http://w3id.org/skg-if/sandbox/acme/otf___1730027051396___product-1
        identifiers:
          type: [array, "null"]
          items:
            required: [ "scheme", "value" ]
            properties:
              scheme:
                type: string
                # alias list in https://skg-if.github.io/context/ver/current/skg-if.json
                # enum: ["doi","pmid","pmcid","arxiv","url","omid","crossref","openalex","orcid","viaf","ror": ,"issn": , "eissn":, "lissn","isbn","opendoar","re3data.org","fairsharing","handle"]
              value:
                type: string
        entity_type:
          type: string
    MetaSingleEntity:
      type: object
      required: [ "local_identifier", "entity_type" ]
      properties:
        local_identifier:
          type: string
          description: Entity API URL. local_identifier can be expressed with ot without full domain. Full domain, including your @base, MUST also resolve
          examples:
            - https://example.com/skg-if/api/products/product-1
            - https://example.com/skg-if/api/persons/person-1
            - https://acme.com/skg-if/api/products/https://w3id.org/skg-if/sandbox/acme/prod-1
            - https://acme.com/skg-if/api/persons/https://w3id.org/skg-if/sandbox/acme/person-1
        entity_type:
          type: string
          description: single entity
          enum:
            - single_entity
        api_items:
          type: [array, "null"]
          items:
            $ref: '#/components/schemas/ApiItem'
    MetaSearch:
      type: object
      required: [ "local_identifier", "entity_type" ]
      properties:
        local_identifier:
          type: string
          description: search result page id as SKG-IF API URL
          examples:
            - https://example.com/skg-if/api/products?filter=xxx&page=y
            - https://example.com/skg-if/api/persons?filter=xxx&page=y
        entity_type:
          type: string
          description: search result page type
          enum:
            - search_result_page
        prev_page:
          description: Previous page reference
          anyOf:
            - $ref: '#/components/schemas/SearchResultPage'
            - type: "null"
        next_page:
          description: Next page reference
          anyOf:
            - $ref: '#/components/schemas/SearchResultPage'
            - type: "null"
        part_of:
          description: Reference for the search result as a whole (all pages)
          type: [object, "null"]
          required: [ "local_identifier", "entity_type" ]
          properties:
            local_identifier:
              type: string
              description: search result id as SKG-IF API URL
              examples:
                - https://example.com/skg-if/api/products?filter=xxx
                - https://example.com/skg-if/api/persons?filter=xxx
            entity_type:
              type: string
              description: search type. Value must be `search_result`
              enum:
                - search_result
            total_items:
              description: total number of item entities for the search
              type: [integer, "null"]
              x-faker:
                datatype.number: [{min: 10, max: 100}]
            first_page:
              description: First page reference of the search
              anyOf:
                - $ref: '#/components/schemas/SearchResultPage'
                - type: "null"
            last_page:
              description: Last page reference of the search
              anyOf:
                - $ref: '#/components/schemas/SearchResultPage'
                - type: "null"
        api_items:
          type: [array, "null"]
          items:
            $ref: '#/components/schemas/ApiItem'
    SearchResultPage:
        type: object
        properties:
          local_identifier:
                type: [string, "null"]
                description: search result page id as SKG-IF API URL
                examples:
                  - https://example.com/skg-if/api/products?filter=xxx&page=y
                  - https://example.com/skg-if/api/persons?filter=xxx&page=y
          entity_type:
            type: [string, "null"]
            description: search result page type. Value must be `search_result_page`
            enum:
              - search_result_page
...
    JsonLdCtxDataModel:
      type: string
      description: "URL to SKG-IF data model context supported by the current OpenAPI `https://w3id.org/skg-if/context/1.1.0/skg-if.json`"
      enum:
        - https://w3id.org/skg-if/context/1.1.0/skg-if.json
      x-faker:
        fake: ['https://w3id.org/skg-if/context/1.1.0/skg-if.json']
    JsonLdCtxSearch:
      type: string
      description: "URL to SKG-IF API context (meta section) supported by the current OpenAPI `https://w3id.org/skg-if/context/1.0.0/skg-if-api.json`"
      enum:
        - https://w3id.org/skg-if/context/1.0.0/skg-if-api.json
      x-faker:
        fake: ['https://w3id.org/skg-if/context/1.0.0/skg-if-api.json']
    JsonLdCtxBaseOrMore:
      description : 'Additional context mappings ex: `{ "@base":"https://w3id.org/skg-if/sandbox/acme" }`'
      properties:
        "@base":
          type: [string, "null"]
          description : |
            Default root URL of the JSON-LD context. \
            To define your `@base`, please refer to the [SKG-IF API doc](https://skg-if.github.io/api/) \
            Your `@base` should end with a `/` \
            common formats : \
            - `https://acme.com/skg/` \
            - `https://w3id.org/acme/` \
            - `https://w3id.org/skg-if/sandbox/acme/`\
          x-faker:
            fake: ['https://w3id.org/skg-if/sandbox/{{random.alpha(5)}}/']
          examples:
            - "https://acme.com/skg/"
            - "https://acme.com/graph/"
            - "https://w3id.org/acme/"
            - "https://w3id.org/skg-if/sandbox/acme/"
      additionalProperties:
        type: string
        description: "context URL of extension entities"
        x-faker:
          fake : ['http://www.example.com/skg/{{random.alpha(3)}}']